Move PReLU to links
