+++ tests/chainer_tests/links_tests/activation_tests/test_prelu.py
